+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changes for croud
 Unreleased
 ==========
 
+- Added support for organization secrets of type Azure.
+
+- Added support for importing jobs from Azure Blob Storage.
+
 1.9.0 - 2023/12/04
 ==================